+++ graphql/mod.ts
@@ -2,13 +2,14 @@ import { type GraphQLSchema, printSchema } from "graphql";
 import path from "node:path";
 import "./account.ts";
 import "./actor.ts";
-import "./reactable.ts";
-import "./post.ts";
-import "./poll.ts";
 import { builder } from "./builder.ts";
+import "./poll.ts";
+import "./post.ts";
+import "./reactable.ts";
+export type { Context } from "./builder.ts";
+export { createYogaServer } from "./server.ts";
 
 export const schema: GraphQLSchema = builder.toSchema();
-export { createYogaServer } from "./server.ts";
 
 void Deno.writeTextFile(
   path.join(import.meta.dirname ?? "", "schema.graphql"),
