+++ models/article.ts
@@ -2,11 +2,11 @@ import type { Context } from "@fedify/fedify";
 import * as vocab from "@fedify/fedify/vocab";
 import { summarize } from "@hackerspub/ai/summary";
 import { translate } from "@hackerspub/ai/translate";
+import { getArticle } from "@hackerspub/federation/objects";
 import { getLogger } from "@logtape/logtape";
 import { minBy } from "@std/collections/min-by";
 import type { LanguageModelV1 } from "ai";
 import { and, eq, isNull, lt, or, sql } from "drizzle-orm";
-import { getArticle } from "../federation/objects.ts";
 import type { ContextData, Models } from "./context.ts";
 import type { Database } from "./db.ts";
 import { syncPostFromArticleSource } from "./post.ts";
