+++ web/routes/@[username]/[idOrYear]/[slug]/index.tsx
@@ -24,13 +24,13 @@ import type {
 import type { Uuid } from "@hackerspub/models/uuid";
 import * as v from "@valibot/valibot";
 import { sql } from "drizzle-orm";
-import { withTransaction } from "../../../../../web/federation.ts";
 import { summarizer } from "../../../../ai.ts";
 import { Msg } from "../../../../components/Msg.tsx";
 import { PageTitle } from "../../../../components/PageTitle.tsx";
 import { PostExcerpt } from "../../../../components/PostExcerpt.tsx";
 import { db } from "../../../../db.ts";
 import { drive } from "../../../../drive.ts";
+import { withTransaction } from "../../../../federation.ts";
 import { ArticleMetadata } from "../../../../islands/ArticleMetadata.tsx";
 import { Composer } from "../../../../islands/Composer.tsx";
 import { PostControls } from "../../../../islands/PostControls.tsx";
