+++ web/routes/@[username]/[idOrYear]/index.tsx
@@ -24,11 +24,11 @@ import type {
 import { type Uuid, validateUuid } from "@hackerspub/models/uuid";
 import * as v from "@valibot/valibot";
 import { sql } from "drizzle-orm";
-import { withTransaction } from "../../../../web/federation.ts";
 import { Msg } from "../../../components/Msg.tsx";
 import { NoteExcerpt } from "../../../components/NoteExcerpt.tsx";
 import { PostExcerpt } from "../../../components/PostExcerpt.tsx";
 import { db } from "../../../db.ts";
+import { withTransaction } from "../../../federation.ts";
 import { Composer } from "../../../islands/Composer.tsx";
 import { PostControls } from "../../../islands/PostControls.tsx";
 import { kv } from "../../../kv.ts";
