+++ web/routes/@[username]/[idOrYear]/quotes.tsx
@@ -18,10 +18,10 @@ import type {
 import { validateUuid } from "@hackerspub/models/uuid";
 import * as v from "@valibot/valibot";
 import { sql } from "drizzle-orm";
-import { withTransaction } from "../../../../web/federation.ts";
 import { Msg } from "../../../components/Msg.tsx";
 import { PostExcerpt } from "../../../components/PostExcerpt.tsx";
 import { db } from "../../../db.ts";
+import { withTransaction } from "../../../federation.ts";
 import { Composer } from "../../../islands/Composer.tsx";
 import { PostControls } from "../../../islands/PostControls.tsx";
 import { define } from "../../../utils.ts";
